adhere return type with the original lib

+++ src/mocks/MockUniswapOracle.sol
@@ -7,7 +7,7 @@ contract MockUniswapOracle is IUniswapOracle
 {
     function getPrice(address token0, address token1, uint24 fee, uint32 twapPeriod)
     external pure
-    returns (uint64 price, uint256 lastUpdateTimestamp)
+    returns (uint128 price, uint256 lastUpdateTimestamp)
     {
         // ETH / USDT
         return (350050, 1);
